Add more detailed, step-by-step integration docs.
As I integrated horus into my project, I wrote what I did in a
step-by-step fashion. Here it is, in reStructuredText format.

It can be easily converted to HTML with *pandoc*:

    $ pandoc -s -t html integration.rst > integration.html

It contains portions of the current README.md. I suppose it could be a
replacement, if renamed to README.rst.

If you really insist on using Markdown, I assume *pandoc* can convert it to
Markdown, too.